Add files via upload
Logic goes here.
The quotes are given in an array.
A random generator selects a number in random and gives it to the other function.
The other function display the Quote present in that particular array's index.
Inside the array, the data is stored as objects i.e. (Key,Value pairs).